move filter popup logic to Filter component
