+++ public/index.html
@@ -16,12 +16,5 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <div class="filter__popup">
-      <div class="popup__text">текст</div>
-      <div class="popup__text">Длинный текст</div>
-      <div class="popup__text">Длинный-длинный текст</div>
-      <div class="popup__text">Длинный-длинный текст</div>
-      <div class="popup__text">Длинный-длинный текст</div>
-    </div>
   </body>
 </html>
