+++ src/authorsData.json
@@ -1,0 +1,22 @@
+[
+    {
+        "id": 0,
+        "author": "Michael Jackson"
+    },
+    {
+        "id": 1,
+        "author": "Frank Sinatra"
+    },
+    {
+        "id": 2,
+        "author": "Arctic Monkeys"
+    },
+    {
+        "id": 3,
+        "author": "Calvin Harris"
+    },
+    {
+        "id": 4,
+        "author": "Zhu"
+    }
+]
