+++ src/genresData.json
@@ -1,0 +1,22 @@
+[
+    {
+        "id": 0,
+        "genre": "Рок"
+    },
+    {
+        "id": 1,
+        "genre": "Хип-хоп"
+    },
+    {
+        "id": 2,
+        "genre": "Поп-музыка"
+    },
+    {
+        "id": 3,
+        "genre": "Техно"
+    },
+    {
+        "id": 4,
+        "genre": "Инди"
+    }
+]
