+++ src/yearsData.json
@@ -1,0 +1,22 @@
+[
+    {
+        "id": 0,
+        "year": "2022"
+    },
+    {
+        "id": 1,
+        "year": "2021"
+    },
+    {
+        "id": 2,
+        "year": "2020"
+    },
+    {
+        "id": 3,
+        "year": "2019"
+    },
+    {
+        "id": 4,
+        "year": "2018"
+    }
+]
